Add files via upload
Scenario: Your company's users create or edit a collection of text files
throughout the day. These text files represent data about customer
orders.
Once per day, any files that are new, or that were edited within the
previous 24 hours, must be sent to the home office. To facilitate this,
these new or updated files need to be copied to a specific 'destination'
folder on a computer, so that a special file transfer program can grab
them and transfer them to the home office.
The process of figuring out which files are new or recently edited, and
copying them to the 'destination' folder, is currently being done
manually. This is very expensive in terms of manpower.
You have been asked to create a script that will automate this task,
saving the company a lot of money over the long term.